% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatialvignette-package.R
\docType{package}
\name{geospatialvignette-package}
\alias{geospatialvignette}
\alias{geospatialvignette-package}
\title{geospatialvignette: Geospatital Vignette Package}
\description{
This package uses packages {sf} {sp} {adehabitatHR} and {fractaldim} to read in, process, analyze and visualize GPS track data.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
