% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mcp_map}
\alias{mcp_map}
\title{mcp_map}
\usage{
mcp_map(sp, perc = 95, units = "km2")
}
\arguments{
\item{sp}{A SpatialPointsDataFrame {sp} containing track (relocation) points of groups or individuals. It should contain 3 columns: "id", "x", "y", where x and y are UTM point coordinates.}

\item{perc}{Percentage of points out of 100 to include in the MCP. Default is 95.}

\item{units}{Units to output area. Takes "km2", "m2", "ha" for hectares. Default is "km2".}
}
\description{
Maps minimum convex polygons around a set of geographic points for one or more groups.
}
\examples{
mcp_map(sp = trackpoints, perc = 90, units = "m2")
}
\keyword{MCP}
\keyword{map}
